new beta version with coverageLogger moved to root level
