+++ src/coverage-reporter/index.ts
@@ -1,5 +1,3 @@
 import GraphqlCoverageReport from './report';
 
 export default GraphqlCoverageReport;
-
-export { coverageLogger } from './coverageLogger';
