+++ tsconfig.json
@@ -70,7 +70,7 @@
     "forceConsistentCasingInFileNames": true        /* Disallow inconsistently-cased references to the same file. */
   },
   "include": [
-    "src/*",
+    "src/**",
     "src/coverage-reporter/*",
     "src/codegen-cli/*"
   ]
